[FD-195] refactor: 모달관련 함수, 컴포넌트 파일 분리
